refactor: moved utils to scripts folder
