+++ scripts/generate-css/styles.ts
@@ -1,5 +1,6 @@
 import { writeFile } from "fs/promises"
 import type { ThemeToken } from "@yamada-ui/react"
+import { prettier, toKebabCase } from "../utils"
 import { getConfig } from "./config"
 import { layoutStyleProperties } from "./layout-props"
 import { resolveTypes } from "./resolve-types"
@@ -9,7 +10,6 @@ import type { TransformOptions } from "./transform-props"
 import { transformMap } from "./transform-props"
 import type { UIOptions } from "./ui-props"
 import { uiProps } from "./ui-props"
-import { prettier, toKebabCase } from "./utils"
 import { OUT_PATH } from "."
 import type { CSSProperties, CSSProperty, UIProperties } from "."
