+++ scripts/utils.ts
